+++ src/components/App/Apps/Messenger/Panels/LeftPanel/Chats.tsx
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+import Header from "./Header";
+
+const StyledChats = styled.div`
+  height: 100%;
+  width: 100%;
+`
+
+const Chats = () => {
+    return (
+        <StyledChats>
+            <Header/>
+        </StyledChats>
+    )
+}
+
+export default Chats